moved social-auth component into shared folder
